chore: corrected folder structure to obey to app router
