+++ next.config.js
@@ -1,8 +1,8 @@
-import createNextIntlPlugin from "next-intl/plugin";
+const createNextIntlPlugin = require("next-intl/plugin");
 
 const withNextIntl = createNextIntlPlugin();
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {};
 
-export default withNextIntl(nextConfig);
+module.exports = withNextIntl(nextConfig);
