+++ src/app/[locale]/layout.tsx
@@ -1,8 +1,10 @@
 import type { Metadata } from "next";
 import { NextIntlClientProvider } from "next-intl";
+import { getMessages } from "next-intl/server";
 import { Montserrat } from "next/font/google";
-import { Toaster } from "./components/ui/toaster";
-import "./globals.css";
+import Navbar from "../components/navbar";
+import { Toaster } from "../components/ui/toaster";
+import "../globals.css";
 import "/node_modules/flag-icons/css/flag-icons.min.css";
 
 const montserrat = Montserrat({ subsets: ["latin"] });
@@ -12,22 +14,27 @@ export const metadata: Metadata = {
   description: "EURO 2024 PICKEMS",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
   params: { locale },
 }: Readonly<{
   children: React.ReactNode;
   params: { locale: string };
 }>) {
-  const messages = require(`../../../messages/${locale}.json`);
+  const messages = await getMessages();
   return (
-    <html lang={"en"} className="dark">
-      <NextIntlClientProvider messages={messages} locale={locale}>
-        <body className={montserrat.className}>
-          {children}
+    <html lang={locale} className="dark">
+      <body className={montserrat.className}>
+        <NextIntlClientProvider messages={messages} locale={locale}>
+          <main className="flex min-h-screen flex-col items-center justify-start bg-black ">
+            <Navbar />
+            <div className="bg-neutral-800 w-11/12 min-h-full flex justify-center items-center p-20 rounded">
+              {children}
+            </div>
+          </main>
           <Toaster />
-        </body>
-      </NextIntlClientProvider>
+        </NextIntlClientProvider>
+      </body>
     </html>
   );
 }
