+++ src/app/[locale]/pickems/page.tsx
@@ -1,15 +1,17 @@
+import { Group } from "@/app/components/group";
+import PointsInfoPopover from "@/app/components/points-info-popover";
+import { Button } from "@/app/components/ui/button";
+import { useToast } from "@/app/components/ui/use-toast";
+import { mockedGroups } from "@/app/mocked-data";
+import { GroupData } from "@/app/types";
+import { useTranslations } from "next-intl";
 import { useState } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { Group } from "../../components/group";
-import PointsInfoPopover from "../../components/points-info-popover";
-import { Button } from "../../components/ui/button";
-import { useToast } from "../../components/ui/use-toast";
-import { mockedGroups } from "../../mocked-data";
-import { GroupData } from "../../types";
 
 const Pickems = (): JSX.Element => {
   const toaster = useToast();
+  const t = useTranslations("PickemsScreen");
 
   const [previousGroups, setPreviousGroups] = useState<GroupData>(
     mockedGroups.data
@@ -40,7 +42,7 @@ const Pickems = (): JSX.Element => {
   return (
     <div className="flex flex-col gap-10 items-center">
       <div className="flex flex-row gap-6">
-        <div className="text-6xl text-green-500">PICKEMS</div>
+        <div className="text-6xl text-green-500">{t("title")}</div>
         <PointsInfoPopover />
       </div>
       <div className="flex flex-wrap flex-row gap-8 justify-center items-center">
@@ -61,7 +63,7 @@ const Pickems = (): JSX.Element => {
         variant="secondary"
         onClick={handleSubmit}
       >
-        Save
+        {t("button")}
       </Button>
     </div>
   );
