+++ src/app/components/group/grab-icon.tsx
