+++ src/middleware.ts
@@ -6,9 +6,10 @@ export default createMiddleware({
 
   // Used when no locale matches
   defaultLocale: "en",
+  localePrefix: "always",
 });
 
 export const config = {
   // Match only internationalized pathnames
-  matcher: ["/", "/(pt|en)/:path*"],
+  matcher: ["/((?!api|_next|...).*)"],
 };
